Make relative font calculations (smaller and larger) more precise. Smaller font for base value 3.75 (i.e. 3.75 / 1.2) did not result in 3.125

+++ src/main/java/com/itextpdf/html2pdf/css/apply/util/FontStyleApplierUtil.java
@@ -225,9 +225,9 @@ public final class FontStyleApplierUtil {
 
     public static float parseRelativeFontSize(final String relativeFontSizeValue, final float baseValue) {
         if (CssConstants.SMALLER.equals(relativeFontSizeValue)) {
-            return baseValue / 1.2f;
+            return (float)(baseValue / 1.2);
         } else if (CssConstants.LARGER.equals(relativeFontSizeValue)) {
-            return baseValue * 1.2f;
+            return (float)(baseValue * 1.2);
         }
         return CssUtils.parseRelativeValue(relativeFontSizeValue, baseValue);
     }
